evaluation updated (#25)
* disable and enable buttons to avoid touching by mistake

* disable and enable buttons to avoid touching by mistake

* name after gesture instead of feature

* change text

* removing png file

* archiving and removing

* add title

* remove png files at evaluation

* return self.archive_csv_name as a static name

* use score_value instead of qualitative_result to evaluate performance

* save evaluation result in my_initializer for later use

* use self.archive_csv_name in Tutorial_Ends_UI

* fix naming convention, space not accepted

* remove_csv_file

* fix return correct archive_csv_name

* != instead of is not

* show_evaluation_result update

* score update

* reduce font scale

* adjust and add UIs

* navigation use target_ui as default

* add Result_StepX_Percent_UIs

* evaluate with calculated differences

* recognized_gesture = 1,2,3

+++ worker_recorder.py
@@ -9,6 +9,7 @@ class WorkerRecorder(QObject):
     permission_to_write = False
 
     current_csv_name = "records/record_current.csv"
+    archive_csv_name = "records/record_archived.csv"
     file_writer = None
 
     @pyqtSlot()
@@ -36,19 +37,15 @@ class WorkerRecorder(QObject):
         self.disable_writing()
         # close file writer before archiving
         self.close_file()
-        try:
-            # archive - renaming file
+        if os.path.exists(self.current_csv_name):
+            # archive - renaming csv file
             archive_time = datetime.datetime.now().strftime("%y%m%d%H%M%S")
             archive_csv_name = "records/record_{}.csv".format(archive_time)
             os.rename(self.current_csv_name, archive_csv_name)
             print("archived successfully")
-            # send signal to create new
-            self.archive_finished.emit()
-        except FileNotFoundError:
-            print("warning - csv file not found")
-            archive_csv_name = None
-            # send signal to create new
-            self.archive_finished.emit()
+        else:
+            print("csv file does not exist")
+        self.archive_finished.emit()
         return archive_csv_name
 
     @pyqtSlot()
@@ -60,13 +57,13 @@ class WorkerRecorder(QObject):
             print("open file writer successfully")
 
     @pyqtSlot()
-    def write_record(self, current_step, sensor_type, result_feature):
+    def write_record(self, current_step, sensor_type, result_gesture):
         if self.permission_to_write and self.file_writer:
             # get current time
             current_time = datetime.datetime.now()
             current_time_s = current_time.strftime("%y%m%d%H%M%S")
             current_time_ms = "{:06d}".format(current_time.microsecond)
             # write into csv file
-            self.file_writer.write("{}{},step_{},{},{}\n".format(current_time_s, current_time_ms, current_step, sensor_type, result_feature))
+            self.file_writer.write("{}{},step_{},{},{}\n".format(current_time_s, current_time_ms, current_step, sensor_type, result_gesture))
         else:
             print("no permission to write")
